Moved boto3 related methods out of CloudWanderer into boto3 interface
