+++ tests/integration/storage_read_generic.py
@@ -1,5 +1,6 @@
 from itertools import combinations
 from time import sleep
+from unittest.mock import patch
 import logging
 from .helpers import TestStorageConnectorReadMixin, setup_moto
 from .mocks import add_infra, generate_mock_session
@@ -19,14 +20,18 @@ class StorageReadTestMixin(TestStorageConnectorReadMixin):
         self.connector = self.connector_class()
         self.connector.init()
         self.memory_storage_connector = cloudwanderer.storage_connectors.MemoryStorageConnector()
-        self.wanderer = cloudwanderer.CloudWanderer(
-            boto3_session=generate_mock_session(),
-            storage_connectors=[self.connector, self.memory_storage_connector]
-        )
-        self.wanderer._account_id = '111111111111'
-        self.wanderer.write_resources()
-        self.wanderer._account_id = '222222222222'
-        self.wanderer.write_resources()
+        with patch('moto.sts.responses.ACCOUNT_ID', new='111111111111'):
+            self.wanderer = cloudwanderer.CloudWanderer(
+                boto3_session=generate_mock_session(),
+                storage_connectors=[self.connector, self.memory_storage_connector]
+            )
+            self.wanderer.write_resources()
+        with patch('moto.sts.responses.ACCOUNT_ID', new='222222222222'):
+            self.wanderer = cloudwanderer.CloudWanderer(
+                boto3_session=generate_mock_session(),
+                storage_connectors=[self.connector, self.memory_storage_connector]
+            )
+            self.wanderer.write_resources()
         self.expected_urns = []
         self.not_expected_urns = []
         # Occasionally moto needs a second to finish writing.
